+++ dwds/test/instances/common/instance_inspection_common.dart
@@ -61,7 +61,7 @@ void runTests({
             verboseCompiler: debug,
             experiments: ['records'],
           ),
-          loadStrategySettings: TestLoadStrategySettings(
+          loadStrategySettings: TestBuildSettings(
             canaryFeatures: canaryFeatures,
           ),
         );