+++ dwds/test/instances/common/patterns_inspection_common.dart
@@ -54,7 +54,7 @@ void runTests({
           verboseCompiler: debug,
           experiments: ['records', 'patterns'],
         ),
-        loadStrategySettings: TestLoadStrategySettings(
+        loadStrategySettings: TestBuildSettings(
           canaryFeatures: canaryFeatures,
         ),
       );