+++ dwds/test/instances/common/record_type_inspection_common.dart
@@ -56,7 +56,7 @@ void runTests({
           verboseCompiler: debug,
           experiments: ['records', 'patterns'],
         ),
-        loadStrategySettings: TestLoadStrategySettings(
+        loadStrategySettings: TestBuildSettings(
           canaryFeatures: canaryFeatures,
         ),
       );